auth_service/requirements.txt
fastapi==0.109.0
uvicorn==0.29.0
sqlalchemy==1.4.40
psycopg2-binary==2.9.7
pydantic==1.10.12
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
kafka-python==2.0.2